Version-specific constructors
